+++ inst/doc/canonical-replications.R
@@ -7,7 +7,8 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>", eval = FALSE)
 # 
 # pc_panel <- ato_individuals_postcode(
 #   year = c("2015-16", "2016-17", "2017-18", "2018-19",
-#            "2019-20", "2020-21", "2021-22", "2022-23")
+#            "2019-20", "2020-21", "2021-22", "2022-23",
+#            "2023-24")
 # )
 # pc_panel <- ato_harmonise(pc_panel)
 #